Import music from any view
