--- src/renderer/components/SettingDropzone/SettingDropzone.module.css
@@ -1,28 +1,0 @@
-.dropzone {
-  border: dashed 1px var(--text);
-  margin-bottom: 15px;
-  padding: 20px;
-  text-align: center;
-  cursor: pointer;
-
-  &:focus {
-    outline: 0;
-  }
-
-  &:active {
-    border-color: var(--main-color);
-  }
-}
-
-.dropzoneDraggedOver {
-  border-color: var(--main-color);
-}
-
-.dropzone__title {
-  font-size: 16px;
-  margin-bottom: 5px;
-}
-
-.dropzone__subtitle {
-  color: var(--text);
-}
