+++ src/renderer/views/Settings/SettingsLibrary.tsx
@@ -2,7 +2,6 @@ import React, { useCallback } from 'react';
 import { ipcRenderer } from 'electron';
 
 import * as Setting from '../../components/Setting/Setting';
-import Dropzone from '../../components/SettingDropzone/SettingDropzone';
 
 import * as LibraryActions from '../../store/actions/LibraryActions';
 import * as PlayerActions from '../../store/actions/PlayerActions';
@@ -15,22 +14,6 @@ interface Props {
 }
 
 const SettingsLibrary: React.FC<Props> = (props) => {
-  const onDrop = useCallback((e: DragEvent) => {
-    const files = [];
-
-    if (e.dataTransfer) {
-      const eventFiles = e.dataTransfer.files;
-
-      for (let i = 0; i < eventFiles.length; i++) {
-        files.push(eventFiles[i].path);
-      }
-
-      LibraryActions.add(files).catch((err) => {
-        console.warn(err);
-      });
-    }
-  }, []);
-
   const resetLibrary = useCallback(async () => {
     PlayerActions.stop();
     await LibraryActions.reset();
@@ -54,15 +37,12 @@ const SettingsLibrary: React.FC<Props> = (props) => {
     <div className='setting settings-musicfolder'>
       <Setting.Section>
         <h3 style={{ marginTop: 0 }}>Import music</h3>
-        <Dropzone
-          title='Add music to the library'
-          subtitle='Click here or drop files or folders anywhere'
-          onDrop={onDrop}
-          onClick={openFolderSelector}
-        />
         <Setting.Description>
           This will also scan for <code>.m3u</code> files and create corresponding playlists.
         </Setting.Description>
+        <Button disabled={props.library.refreshing} onClick={openFolderSelector}>
+          Add files or folders
+        </Button>
       </Setting.Section>
       <Setting.Section>
         <h3>Danger zone</h3>
