+++ src/themes/dark-legacy.json
@@ -5,6 +5,8 @@
   "variables": {
     "--text": "#EEEEEE",
     "--background": "#23292B",
+    "--faded": "rgba(0, 0, 0, 0.8)",
+
     "--border-color": "#1F1F1F",
     "--border-radius": "2px",
