+++ src/themes/light.json
@@ -5,6 +5,8 @@
   "variables": {
     "--text": "#333333",
     "--background": "white",
+    "--faded": "rgba(255, 255, 255, 0.9)",
+
     "--border-color": "#D2D2D2",
     "--border-radius": "2px",
